// generated/src/app/api-models/prov-cross-ref.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class ProvCrossRef {
  seqProvxId: number;
  payorId: string;
  payorCode: string;
  seqProvId: number;
  seqProvAddress: number;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}